+++ src/components/Header/HeaderLogo.jsx
@@ -1,0 +1,9 @@
+const HeaderLogo = () => {
+    return (
+        <a href="#" className="logo-padding p-">
+            <img src="/yt-logo.svg" draggable="false" alt="로고"></img>
+        </a>
+    );
+};
+
+export default HeaderLogo;
