+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import Header from '../components/Header';
+import Header from '../components/Header/Header';
 import Sidebar from '../components/Sidebar';
 import VideoCard from '../components/VideoCard';
